+++ pkg/deployment/member/phase_updates.go
@@ -100,6 +100,7 @@ func removeMemberConditionsMapFunc(m *api.MemberStatus) {
 	m.Conditions.Remove(api.ConditionTypeTopologyAware)
 	m.Conditions.Remove(api.MemberReplacementRequired)
 	m.Conditions.Remove(api.ConditionTypePVCResizePending)
+	m.Conditions.Remove(api.ConditionTypeScaleDownCandidate)
 
 	m.RemoveTerminationsBefore(time.Now().Add(-1 * recentTerminationsKeepPeriod))