+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - (Feature) Propagate sidecars' ports to a member's service
 - (Debug Package) Initial commit
 - (Feature) Detach PVC from deployment in Ordered indexing method
+- (Feature) ScaleDown Candidate
 
 ## [1.2.16](https://github.com/arangodb/kube-arangodb/tree/1.2.16) (2022-09-14)
 - (Feature) Add ArangoDeployment ServerGroupStatus